requirements.txt
# ======================================================
# Core Deep Learning
# ======================================================
torch
torchvision
torchaudio

# ======================================================
# HuggingFace Transformers (SegFormer)
# ======================================================
transformers
huggingface-hub
safetensors

# ======================================================
# FastAPI Backend
# ======================================================
fastapi
uvicorn
python-multipart

# ======================================================
# Image Processing
# ======================================================
pillow
opencv-python

# ======================================================
# Data Processing
# ======================================================
numpy
scipy

# ======================================================
# Visualization / Debugging
# ======================================================
matplotlib
tqdm

# ======================================================
# Utility Libraries
# ======================================================
requests
pydantic
typing-extensions